requirements.txt
tensorflow-gpu==2.6.0
tensorflow-probability
tensorflow-datasets

hydra-core
hydra_colorlog
omegaconf
cloud-tpu-client
rich